requirements.txt
requests
PyYAML
tqdm
pip-audit
inquirer